Refactor : ChannelList의 ChannelListHeader 부분을 별도의 컴포넌트로 분리
