+++ frontend/src/components/Main/ChannelList.tsx
@@ -5,10 +5,9 @@ import styled from 'styled-components';
 import { BoostCamMainIcons } from '../../utils/SvgIcons';
 import { ChannelData } from '../../types/main';
 import { MainStoreContext } from './MainStore';
-import Dropdown from '../core/Dropdown';
-import DropdownMenu from '../core/DropdownMenu';
+import ChannelListHeader from './ChannelListHeader';
 
-const { Hash, Plus, ListArrow } = BoostCamMainIcons;
+const { Hash } = BoostCamMainIcons;
 
 const Container = styled.div`
   width: 100%;
@@ -20,33 +19,6 @@ const Container = styled.div`
   flex-direction: column;
   justify-content: flex-start;
   align-items: flex-start;
-`;
-
-const ChannelListHeader = styled.div`
-  width: 90%;
-  height: 30px;
-
-  margin-left: 15px;
-  color: #a69c96;
-  font-size: 17px;
-
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: center;
-
-  &:hover {
-    cursor: pointer;
-  }
-`;
-
-const ChannelListHeaderSpan = styled.span`
-  margin-left: 5px;
-`;
-
-const ChannelListHeaderButton = styled.div<{ isButtonVisible: boolean }>`
-  margin-left: 70px;
-  visibility: ${(props) => (props.isButtonVisible ? 'visible' : 'hidden')};
 `;
 
 const ChannelListBody = styled.div`
@@ -83,20 +55,6 @@ const ChannelNameSpan = styled.span`
   padding: 5px 0px 5px 5px;
 `;
 
-const ListArrowIcon = styled(ListArrow)<{ isListOpen: boolean }>`
-  width: 20px;
-  height: 20px;
-  fill: #a69c96;
-  transition: all ease-out 0.3s;
-  ${(props) => (props.isListOpen ? 'transform: rotate(90deg);' : 'transform: rotate(0deg);')}
-`;
-
-const PlusIcon = styled(Plus)`
-  width: 20px;
-  height: 20px;
-  fill: #a69c96;
-`;
-
 const HashIcon = styled(Hash)`
   width: 15px;
   height: 15px;
@@ -104,29 +62,13 @@ const HashIcon = styled(Hash)`
 `;
 
 function ChannelList(): JSX.Element {
-  const [isButtonVisible, setIsButtonVisible] = useState<boolean>(false);
-  const [isDropdownActivated, setIsDropdownActivated] = useState<boolean>(false);
   const [isListOpen, setIsListOpen] = useState<boolean>(false);
-  const {
-    selectedServer,
-    selectedChannel,
-    serverChannelList,
-    isCreateModalOpen,
-    isJoinModalOpen,
-    setSelectedChannel,
-    setIsCreateModalOpen,
-    setIsJoinModalOpen,
-  } = useContext(MainStoreContext);
+  const { selectedServer, selectedChannel, serverChannelList, setSelectedChannel } = useContext(MainStoreContext);
   const navigate = useNavigate();
 
   const onClickChannelBlock = ({ currentTarget }: React.MouseEvent<HTMLDivElement>) => {
     const channelId = currentTarget.dataset.id;
     if (channelId) setSelectedChannel(channelId);
-  };
-
-  const onClickChannelAddButton = (e: React.MouseEvent<HTMLOrSVGElement>) => {
-    e.stopPropagation();
-    setIsDropdownActivated(!isDropdownActivated);
   };
 
   const onRightClickChannelItem = (e: React.MouseEvent<HTMLDivElement>) => {
@@ -162,31 +104,7 @@ function ChannelList(): JSX.Element {
 
   return (
     <Container>
-      <ChannelListHeader
-        onMouseEnter={() => setIsButtonVisible(true)}
-        onMouseLeave={() => setIsButtonVisible(false)}
-        onClick={() => setIsListOpen(!isListOpen)}
-      >
-        <ListArrowIcon isListOpen={isListOpen} />
-        <ChannelListHeaderSpan>채널</ChannelListHeaderSpan>
-        <ChannelListHeaderButton isButtonVisible={isButtonVisible}>
-          <PlusIcon onClick={onClickChannelAddButton} />
-          <Dropdown isDropdownActivated={isDropdownActivated} setIsDropdownActivated={setIsDropdownActivated}>
-            <DropdownMenu
-              name="추가"
-              setIsDropdownActivated={setIsDropdownActivated}
-              state={isCreateModalOpen}
-              stateSetter={setIsCreateModalOpen}
-            />
-            <DropdownMenu
-              name="생성"
-              setIsDropdownActivated={setIsDropdownActivated}
-              state={isJoinModalOpen}
-              stateSetter={setIsJoinModalOpen}
-            />
-          </Dropdown>
-        </ChannelListHeaderButton>
-      </ChannelListHeader>
+      <ChannelListHeader isListOpen={isListOpen} setIsListOpen={setIsListOpen} />
       {isListOpen && <ChannelListBody>{listElements}</ChannelListBody>}
     </Container>
   );
